\name{newtoncooling}
\alias{newtoncooling}
\title{newtoncooling}
\description{NewtonColling}
\usage{newtoncooling(a1, a2)}
\arguments{
  \item{a1}{coefficient, see details}
  \item{a2}{coefficient, see details}
}
\details{derivate function for system
\cr dx = a1*(y-x)
\cr dy = a2*(x-y)}
\value{closure calculating the derivative based on vector arguments (x,y).}

\author{Thomas Wutzler}





