\name{phaseContours}
\alias{phaseContours}
\title{phaseContours}
\description{draw contour lines for the y and x component of the derivative}
\usage{phaseContours(z, xy, add = F, colors = c("red", "blue"), ...)}
\arguments{
  \item{z}{result component of \code{\link{evalDerivGrid}}}
  \item{xy}{result component of \code{\link{evalDerivGrid}}}
  \item{add}{set to TRUE to add arrows to an existing plot}
  \item{colors}{colors of the contourlines for x and y direction of the derivative}
  \item{\dots}{further arguments to \code{\link{contour}}, such as \code{levels}}
}



\author{Thomas Wutzler}





