\name{phaseNullclines}
\alias{phaseNullclines}
\title{phaseNullclines}
\description{draw nullclines, estimated by contour line at level 0}
\usage{phaseNullclines(...)}
\arguments{
  \item{\dots}{arguments to \code{\link{phaseContours}}}
}



\author{Thomas Wutzler}





